Added!
An app created in half a day. This project utilizes various technologies including React as the framework and TypeScript as the primary language. For the API server, Express.js and Node.js are employed, with SQL databases for data storage.
Additionally, Tailwind CSS with the RippleUI plugin is used for styling.